requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Deprecated==1.2.13
Flask==2.0.2
gunicorn==20.1.0
humanize==3.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
MarkupSafe==2.0.1
mwparallelparser==0.1.6
mysql-connector-python==8.0.27
nltk==3.6.5
numpy==1.22.2
pandas==1.4.1
protobuf==3.19.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.0.2
regex==2021.10.23
requests==2.26.0
six==1.16.0
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
